Refactoring : renomage de l'extention vue texte_a_trous.vue en liste_courses_a_trous.vue
