+++ src/situations/cafe_de_la_place/vues/cafe_de_la_place.vue
@@ -25,7 +25,7 @@ import TransitionFade from 'commun/vues/transition_fade';
 import Pagination from 'commun/vues/components/pagination';
 
 import ClicSurMots from 'cafe_de_la_place/vues/components/clic_sur_mots.vue';
-import texteATrous from 'cafe_de_la_place/vues/components/texte_a_trous.vue';
+import listeCoursesATrous from 'cafe_de_la_place/vues/components/liste_courses_a_trous.vue';
 import Puzzle from 'cafe_de_la_place/vues/components/puzzle.vue';
 import Graphique from 'cafe_de_la_place/vues/components/graphique.vue';
 
@@ -35,7 +35,7 @@ export default {
   mounted () {
     // enregistre globalement, les composants spécifiques utilisés par certaines questions
     Vue.component('clic_sur_mots', ClicSurMots);
-    Vue.component('texte_a_trous', texteATrous);
+    Vue.component('liste_courses_a_trous', listeCoursesATrous);
     Vue.component('puzzle', Puzzle);
     Vue.component('graphique', Graphique);
   },
