+++ src/situations/cafe_de_la_place/vues/components/liste_courses_a_trous.vue
